requirements.txt
albucore==0.0.23
albumentations==2.0.4
annotated-types==0.7.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
gdown==5.2.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgviz==1.7.6
kiwisolver==1.4.8
labelme==5.6.1
lazy_loader==0.4
loguru==0.7.3
matplotlib==3.10.0
mpmath==1.3.0
natsort==8.4.0
networkx==3.4.2
numpy==2.2.2
onnxruntime==1.20.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
osam==0.2.2
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
QtPy==2.4.3
requests==2.32.3
scikit-image==0.25.1
scipy==1.15.1
simsimd==6.2.1
six==1.17.0
soupsieve==2.6
stringzilla==3.11.3
sympy==1.13.3
termcolor==2.5.0
tifffile==2025.1.10
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
